Multithreading Update!
A very exciting update:  Introduced threading in the O(n^2) parts of the simulation!  Currently Particle-Particle Collisions and Gravity benefit from threading.  The expected gain in number of particles able to simulate in real time is ~50% with a four core processor.

+++ Main.java
@@ -52,8 +52,8 @@ public class Main extends JFrame {
 			//Begin timer
 			long new_frame_time = System.nanoTime();
 			
-			double accuracy_multiple = this.game.accuracy_multiple;
-			double secs_per_sec = this.game.secs_per_sec;
+			int	accuracy_multiple = this.game.accuracy_multiple;
+			int secs_per_sec = this.game.secs_per_sec;
 			if (accuracy_multiple <= 0)
 			{
 				System.out.println("Accuracy must be set to 0 or higher!");
@@ -62,7 +62,7 @@ public class Main extends JFrame {
 			
 			double timestep = 1.0/accuracy_multiple;
 			this.game.timestep = timestep;
-			for (int i = 0; i < accuracy_multiple * secs_per_sec; i++)
+			for (long i = 0; i < accuracy_multiple * secs_per_sec; i++)
 				this.game.update();
 			
 			repaint(); // Indirectly calls View.paintComponent in its own thread (?), will paint whenever it feels like it
